+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Folding@home Client Changelog
  - Log outgoing URLs. @Br3ach
  - Restart tray icon on Windows when tray restarts.  @SortaCore re:#437
  - Use PCI domain when detecting GPUs.  re:#346
+ - Fix upload retry issue.  re:#447
 
 ## v8.5.6
  - Failing ``config.xml`` load logs error but is now non-fatal.